[53806] Port is working w hard-coded network data.
Started adding code to handle TLS connection to RVI server, but for some reason not connecting to end-point. Possibly issue w network or simulator. Bypassed the issue by hard-coding RVI messages into RVIServerConnection so that I could test the rest of the port. HVACManager is configuring and connecting RVI node. Hard-coded service invocation sent, and successfully received. Much success!

+++ RVI/RVINode.h
@@ -55,7 +55,7 @@
  *
  * @param serverPort the server port
  */
-- (void)setServerPort:(NSInteger)serverPort;
+- (void)setServerPort:(UInt32)serverPort;
 
 /**
  * Sets the server port of the remote RVI node, when using a TCP/IP link to interface with a remote node.
