+++ RVI/RVIRemoteConnectionManager.h
@@ -97,7 +97,7 @@ typedef enum
  *
  * @param serverPort the server port
  */
-- (void)setServerPort:(NSInteger)serverPort;
+- (void)setServerPort:(UInt32)serverPort;
 
 /**
  * Sets the trusted server certificate of the remote RVI node, when using a TCP/IP link to interface with a remote node.
